backend/requirements.txt
flask
flask-cors
openai
google-generativeai
chromadb
pillow
requests
python-dotenv
